requirements.txt
# Core Web Framework
flask==2.3.3

# PyTorch
torch==1.11.0
torchvision==0.12.0

# NLP + OCR Models
transformers==4.38.0
sentencepiece==0.1.99
tokenizers==0.15.1
accelerate==0.28.0
easyocr==1.6.2

# CRAFT Text Detection
craft-text-detector==0.3.4

# Image Processing
Pillow==10.2.0
opencv-contrib-python==4.5.5.62

# Metrics & Utilities 
jiwer==3.0.3
editdistance==0.6.2
numpy==1.26.4

# Dataset Loader
datasets==2.18.0
